Allow a curator who deposited a work the ability to edit the work after it has been approved
Moved logic for edit button to the work, so it can be utilized in the controller and the view.  The view logic and the controller logic was not quite complete.  The view logic did not care if the user had edit rights for the item.  The controller logic did not allow administrators to edit the work if they submitted the work.

fixes #833