Re-route the call to always fetch pre-curation files directly from AWS (bypassing ActiveStorage) (#962)
Kept the old pre_curation_uploads property in place and added one pre_curation_uploads_fast that goes straight to AWS

Removed comparison of local files storage with AWS storage on the show page, since everything is now showing what is in S3.

Updating the tests mocking to allow for the access to S3 always.  Before we were utilizing our local copy in many of the test scenarios.