Create a MigratioUploadSnapshot to allow dspace files to be marked as migration
Changed AWS copy to run in single copy if the file was smaller than 5GB to make the MD5 checksum match most of the time.

Add new activity types for migration so we can keep track of the start of the migration, but only show the end on the user interface.

Add renderer for new activity type

Also moved PULDspaceData to PULDspaceMigrate
Separated the DSpaceFiles and DspaceAWSFiles into separate classes to make the main migration class cleaner